tweaks: adjust the type of four lockfile variables
The functions read() and fwrite() take size_t, not ssize_t.

And line numbers in the file should be displayed as a long type instead
of an int, since the effective type of ssize_t is not int, but long.

+++ src/utils.c
@@ -642,8 +642,8 @@ filestruct *fsfromline(ssize_t lineno)
 	    f = f->next;
 
     if (f->lineno != lineno) {
-	statusline(ALERT, _("Internal error: can't match line %d.  "
-			"Please save your work."), lineno);
+	statusline(ALERT, _("Internal error: can't match line %ld.  "
+			"Please save your work."), (long)lineno);
 	return NULL;
     }
